Move pages into ui package, renames, move database files into source package
