+++ app/src/main/java/com/example/animalcount/MainActivity.kt
@@ -3,10 +3,9 @@ package com.example.animalcount
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.example.animalcount.navigation.pages.animallist.AnimalListPage
-import com.example.animalcount.ui.components.AnimalCountScaffold
+import com.example.animalcount.ui.pages.home.HomePage
+import com.example.animalcount.ui.components.AppScaffold
 import com.example.animalcount.ui.theme.AnimalCountTheme
-import com.example.animalcount.util.SampleAnimalList
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -15,8 +14,8 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             AnimalCountTheme {
-                AnimalCountScaffold {
-                    AnimalListPage()
+                AppScaffold {
+                    HomePage()
                 }
             }
         }
